+++ service/controller/key/key.go
@@ -1,9 +1,9 @@
 package key
 
 import (
-	securityv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/security/v1alpha1"
 	"github.com/giantswarm/k8smetadata/pkg/annotation"
 	"github.com/giantswarm/microerror"
+	securityv1alpha1 "github.com/giantswarm/organization-operator/api/v1alpha1"
 )
 
 func ToOrganization(v interface{}) (securityv1alpha1.Organization, error) {